requirements.txt
aiogram==2.5.3
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
Babel==2.15.0
certifi==2024.7.4
charset-normalizer==3.3.2
config==0.5.1
decouple==0.0.7
dj-database-url==2.2.0
Django==4.2.14
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
django-parler
django-parler-rest
environs==11.0.0
frozenlist==1.4.1
gunicorn==22.0.0
idna==3.7
marshmallow==3.21.3
multidict==6.0.5
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
sqlparse==0.5.1
telegram==0.0.1
thread6==0.2.0
typing-extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
whitenoise==6.4.0
yarl==1.9.4
